chore: update submodules and configuration for local development
Add a new submodule "themes/terminal" pointing to the same repository as "themes/hugo-theme-terminal" to facilitate theme management. Update the `hugo.toml` configuration to use the "terminal" theme and change the `baseURL` to "http://localhost:1313/" for local development purposes. Comment out the previous `baseURL` to prevent conflicts during local testing. Remove the footer parameter as it is no longer needed. Delete the `cover.jpg` file from the resources directory as it is not required. These changes streamline the development process by aligning the configuration with local testing needs and removing unnecessary files.